Doors And Interfaces
Changed from using gameobject.sendmessages to giving objects a base template that has a slew of interfaces attached that can be overridden. That means anything that can be picked up can have optional functionality implemented onto it.

So it used getcomponent<IPickUpAction>().PickUpAction(); which I feel is much better than passing a string arguement.

That and doors and more stuff for buttons like giving them sounds for success. More in world feedback.