Seperation of Functionality
Made changes to the computer console system. Seperated all functionality to where they should be. I.e
Don't want to have the system to activate the console be in the same place as validating input and displaying contents.

Now input validation and console output is seperate. This is for ease of readability.

Works now by having Events that are listened for by the ConsoleSystem which in turns builds string and sends those strings to the ConsoleOutput class that handles only the displaying of content.